refactor: simplify module names in population package
