+++ worldometer/world/population/__init__.py
@@ -10,6 +10,6 @@ __all__ = [
 from worldometer.world.population.countries_by_population import CountriesByPopulation
 from worldometer.world.population.largest_cities import LargestCities
 from worldometer.world.population.most_populous_countries import MostPopulousCountries
-from worldometer.world.population.world_population_by_region import WorldPopulationByRegion
-from worldometer.world.population.world_population_by_year import WorldPopulationByYear
-from worldometer.world.population.world_population_projections import WorldPopulationProjections
+from worldometer.world.population.by_region import WorldPopulationByRegion
+from worldometer.world.population.by_year import WorldPopulationByYear
+from worldometer.world.population.projections import WorldPopulationProjections
