+++ worldometer/world/population/by_year.py
